Remove ChecklistManager in favor of FileStore
