+++ hub/commands.go
@@ -11,8 +11,6 @@ import (
 	"golang.org/x/xerrors"
 	"google.golang.org/grpc"
 
-	"github.com/greenplum-db/gpupgrade/hub/upgradestatus"
-	"github.com/greenplum-db/gpupgrade/idl"
 	"github.com/greenplum-db/gpupgrade/utils"
 	"github.com/greenplum-db/gpupgrade/utils/daemon"
 	"github.com/greenplum-db/gpupgrade/utils/log"
@@ -64,28 +62,7 @@ func Command() *cobra.Command {
 				return err
 			}
 
-			cm := upgradestatus.NewChecklistManager(stateDir)
-
-			h := New(conf, grpc.DialContext, stateDir, cm)
-
-			// Set up the checklist steps in order.
-			//
-			// TODO: make sure the implementations here, and the Checklist below, are
-			// fully exercised in end-to-end tests. It feels like we should be able to
-			// pull these into a Hub method or helper function, but currently the
-			// interfaces aren't well componentized.
-			cm.AddWritableStep(upgradestatus.CONFIG, idl.Substep_CONFIG)
-			cm.AddWritableStep(upgradestatus.START_AGENTS, idl.Substep_START_AGENTS)
-			cm.AddWritableStep(upgradestatus.CREATE_TARGET_CONFIG, idl.Substep_CREATE_TARGET_CONFIG)
-			cm.AddWritableStep(upgradestatus.SHUTDOWN_SOURCE_CLUSTER, idl.Substep_SHUTDOWN_SOURCE_CLUSTER)
-			cm.AddWritableStep(upgradestatus.INIT_TARGET_CLUSTER, idl.Substep_INIT_TARGET_CLUSTER)
-			cm.AddWritableStep(upgradestatus.SHUTDOWN_TARGET_CLUSTER, idl.Substep_SHUTDOWN_TARGET_CLUSTER)
-			cm.AddWritableStep(upgradestatus.CHECK_UPGRADE, idl.Substep_CHECK_UPGRADE)
-			cm.AddWritableStep(upgradestatus.UPGRADE_MASTER, idl.Substep_UPGRADE_MASTER)
-			cm.AddWritableStep(upgradestatus.COPY_MASTER, idl.Substep_COPY_MASTER)
-			cm.AddWritableStep(upgradestatus.UPGRADE_PRIMARIES, idl.Substep_UPGRADE_PRIMARIES)
-			cm.AddWritableStep(upgradestatus.START_TARGET_CLUSTER, idl.Substep_START_TARGET_CLUSTER)
-			cm.AddWritableStep(upgradestatus.RECONFIGURE_PORTS, idl.Substep_RECONFIGURE_PORTS)
+			h := New(conf, grpc.DialContext, stateDir)
 
 			if shouldDaemonize {
 				h.MakeDaemon()
