+++ hub/upgrade_master.go
@@ -3,6 +3,7 @@ package hub
 import (
 	"os/exec"
 
+	"github.com/greenplum-db/gpupgrade/step"
 	"github.com/greenplum-db/gpupgrade/upgrade"
 	"github.com/greenplum-db/gpupgrade/utils"
 )
@@ -13,7 +14,7 @@ var execCommand = exec.Command
 // XXX this makes more sense as a Hub method, but it's so difficult to stub a
 // Hub that the parameters have been split out for testing. Revisit if/when the
 // Hub monolith is broken up.
-func UpgradeMaster(source, target *utils.Cluster, stateDir string, stream OutStreams, checkOnly bool, useLinkMode bool) error {
+func UpgradeMaster(source, target *utils.Cluster, stateDir string, stream step.OutStreams, checkOnly bool, useLinkMode bool) error {
 	wd := utils.MasterPGUpgradeDirectory(stateDir)
 	err := utils.System.MkdirAll(wd, 0700)
 	if err != nil {
