--- hub/upgradestatus/upgradestatus_suite_test.go
@@ -1,13 +1,0 @@
-package upgradestatus_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestCommands(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Upgrade Status Suite")
-}
